@@ -3,11 +3,13 @@ package com.example.asimyamin.blood_donation.Adapter;
 import android.content.Context;
 import android.content.Intent;
 import android.net.Uri;
-import android.support.v7.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.TextView;
+
+import androidx.recyclerview.widget.RecyclerView;
+
 import com.example.asimyamin.blood_donation.Model.DonorInfomation;
 import com.example.asimyamin.blood_donation.R;
 
